requirements/local.txt
-r base.txt
django-debug-toolbar==4.4.6
pytest==8.3.3
pytest-django==4.9.0
pytest-cov==6.0.0
junit-xml==1.9